pyproject.toml
[tool.poetry]
name = "requirements"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.3"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

